move where the api request is called from
