+++ packages/studio/src/app/appToolbar.tsx
@@ -1,21 +1,20 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { DataToolbar , DataToolbarItem, DataToolbarContent } from '@patternfly/react-core/dist/esm/experimental';
 import { Button } from '@patternfly/react-core';
 import {ThIcon, ListIcon} from '@patternfly/react-icons';
-import './app.css'
+import './app.css';
+import { StoreContext } from './../context/StoreContext';
+import { useStoreContext } from './../context/reducers';
 
-interface AppToolbarProps {
-  buttonClick: (ev: React.MouseEvent<HTMLButtonElement>) => void,
-  buttonSelected: string
-}
+export const AppToolbar = () => {
 
+  const [state, setState] = useContext(StoreContext);
+  const { dashboardState } = useStoreContext();
 
-export class AppToolbar extends React.Component<AppToolbarProps> {
-  constructor(props: AppToolbarProps) {
-    super(props);
+  const setDashboardState = (dashboardState: string) => {
+    setState({...state, dashboardView: dashboardState});
   }
 
-  render() {
     return (
       <DataToolbar>
         <DataToolbarContent>
@@ -23,10 +22,10 @@ export class AppToolbar extends React.Component<AppToolbarProps> {
             This is where the Data Toolbar should go
           </DataToolbarItem>
           <DataToolbarItem variant="pagination">
-            <Button onClick={this.props.buttonClick} className={'app-data-toolbar-button-control ' + (this.props.buttonSelected === "card" ? "pf-m-selected" : "")} variant="plain">
+            <Button onClick={() => setDashboardState('card')} className={'app-data-toolbar-button-control ' + (dashboardState === "card" ? "pf-m-selected" : "")} variant="plain">
               <ThIcon/>
             </Button>
-            <Button onClick={this.props.buttonClick} className={'app-data-toolbar-button-control ' + (this.props.buttonSelected === "list" ? "pf-m-selected" : "")} variant="plain">
+            <Button onClick={() => setDashboardState('list')} className={'app-data-toolbar-button-control ' + (dashboardState === "list" ? "pf-m-selected" : "")} variant="plain">
               <ListIcon/>
             </Button>
             <span className="app-toolbar-api-total">
@@ -37,6 +36,5 @@ export class AppToolbar extends React.Component<AppToolbarProps> {
       </DataToolbar>
     )
   }
-}
 
 export default AppToolbar;
